Override Utils.user_path to not spawn logs in the AP logs folder
As the comment says, AP starts a thread to cleanup logs when
initializing logging. Which means that it starts that thread for every
fuzz attempt. Except that it scans all log files every time, to delete
the ones that are more than 7 days old. Which is none in our case, we
keep piling logs and making that thread busier and busier to the point
where it starts to impact performance very badly.