Stop waiting for timeouts to clear themselves
Because `generate_output` is threaded and has some barriers, if it times
out while it's in there for a good reason, it ends up deadlocking. So
instead of waiting for threads to process the exception forever, kill
them savagely after giving them some time.

Hopefully that works on windows too...