Print the progress report differently when stdout is not a tty
This is mostly useful for CI as taskcluster won't print anything until
it gets a new line and thus doesn't print anything at all during the
whole fuzzing process which renders livelogs useless and means that I
have to wait until the end of the job to even have an idea of what's
going on.
This will print ~50 lines throughout the fuzzing process which isn't
excessive. We'll see how that goes and hopefully it'll behave on
windows...